LayoutTests/imported/w3c/web-platform-tests/domparsing/interfaces.any.worker-expected.txt


PASS idl_test setup 
PASS idl_test validation 
PASS Partial interface Element: original interface defined 
PASS Partial interface Element: member names are unique 
PASS Partial interface Range: original interface defined 
PASS Partial interface Range: member names are unique 
PASS Element includes InnerHTML: member names are unique 
PASS Element includes ParentNode: member names are unique 
PASS Element includes NonDocumentTypeChildNode: member names are unique 
PASS Element includes ChildNode: member names are unique 
PASS Element includes Slotable: member names are unique 
PASS ShadowRoot includes InnerHTML: member names are unique 
PASS DocumentFragment includes NonElementParentNode: member names are unique 
PASS DocumentFragment includes ParentNode: member names are unique 
PASS DOMParser interface: existence and properties of interface object 
FAIL DOMParser must be primary interface of new DOMParser() assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: DOMParser"
FAIL Stringification of new DOMParser() assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: DOMParser"
FAIL DOMParser interface: new DOMParser() must not have property "undefined" assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: DOMParser"
FAIL DOMParser interface: new DOMParser() must not have property "parseFromString" assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: DOMParser"
PASS XMLSerializer interface: existence and properties of interface object 
FAIL XMLSerializer must be primary interface of new XMLSerializer() assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: XMLSerializer"
FAIL Stringification of new XMLSerializer() assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: XMLSerializer"
FAIL XMLSerializer interface: new XMLSerializer() must not have property "undefined" assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: XMLSerializer"
FAIL XMLSerializer interface: new XMLSerializer() must not have property "serializeToString" assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: XMLSerializer"
PASS Node interface: existence and properties of interface object 
PASS DocumentFragment interface: existence and properties of interface object 
PASS ShadowRoot interface: existence and properties of interface object 
PASS Element interface: existence and properties of interface object 
FAIL Element interface: document.createElement("div") must not have property "outerHTML" assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: document"
FAIL Element interface: document.createElement("div") must not have property "insertAdjacentHTML" assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: document"
FAIL Element interface: document.createElement("div") must not have property "innerHTML" assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: document"
PASS AbstractRange interface: existence and properties of interface object 
PASS Range interface: existence and properties of interface object 
FAIL Range interface: new Range() must not have property "createContextualFragment" assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: Range"
